Rename files and restructure some locations
