+++ src/helpers/inputManager.ts
