+++ src/menu/helpMenu.ts
@@ -1,5 +1,5 @@
 import { shell, MenuItemConstructorOptions } from "electron";
-import { IS_MAC, IS_WINDOWS } from "../constants";
+import { IS_MAC, IS_WINDOWS } from "../helpers/constants";
 import { aboutMenuItem } from "./items/about";
 import { checkForUpdatesMenuItem } from "./items/check_for_updates";
 import { separator } from "./items/separator";
